requirements.txt
asgiref==3.5.2
black==22.8.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.2
django-crispy-forms==1.14.0
django-environ==0.9.0
django-import-export==2.9.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.10
pathspec==0.10.1
platformdirs==2.5.2
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tablib==3.2.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
